requirements.txt
click==8.0.3
Flask==2.0.2
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.14.0
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.2
